Registered the errors blueprint in app/__init__.py and added a condition that aborts requests (/mark, /delete, /update and /view) in the app/task/routes.py if the current_user is not the owner that particular task

+++ app/task/routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, flash, redirect, render_template, request, url_for
+from flask import Blueprint, abort, flash, redirect, render_template, request, url_for
 from flask_login import current_user, login_required
 from sqlalchemy import func
 from app.forms import TaskForm
@@ -33,6 +33,8 @@ def update_task(task_id):
     form = TaskForm()
 
     task = Task.query.get_or_404(int(task_id))
+    if current_user != task.user:
+        abort(403)
     if form.validate_on_submit():
         task.task = form.task.data
         task.deadline = form.deadline.data
@@ -51,6 +53,8 @@ def update_task(task_id):
 @login_required
 def delete_task(task_id):
     task = Task.query.get_or_404(int(task_id))
+    if current_user != task.user:
+        abort(403)
     db.session.delete(task)
     db.session.commit()
 
@@ -62,6 +66,8 @@ def delete_task(task_id):
 @login_required
 def mark_task(task_id):
     task = Task.query.get_or_404(int(task_id))
+    if current_user != task.user:
+        abort(403)
     if task.is_completed == False:
         task.is_completed = True
         task.user.completed_task +=  1       
@@ -74,5 +80,7 @@ def mark_task(task_id):
 @login_required
 def view_task(task_id):
     task = Task.query.get_or_404(int(task_id))
+    if current_user != task.user:
+        abort(403)
     return render_template("task/view_task.html", task=task, title="View Task")
